Move wallet test trait to testkit
