+++ wallet-test/src/test/scala/org/bitcoins/wallet/fixtures/UtxoDAOFixture.scala
@@ -2,7 +2,7 @@ package org.bitcoins.wallet.fixtures
 
 import org.bitcoins.wallet.db.WalletDbManagement
 import org.bitcoins.wallet.models.UTXOSpendingInfoDAO
-import org.bitcoins.wallet.util.BitcoinSWalletTest
+import org.bitcoins.testkit.wallet.BitcoinSWalletTest
 import org.scalatest._
 
 import scala.concurrent.Future
