+++ wallet-test/src/test/scala/org/bitcoins/wallet/models/AccountDAOTest.scala
@@ -2,7 +2,7 @@ package org.bitcoins.wallet.models
 
 import org.bitcoins.testkit.core.gen.CryptoGenerators
 import org.bitcoins.wallet.fixtures.AccountDAOFixture
-import org.bitcoins.wallet.util.{BitcoinSWalletTest, WalletTestUtil}
+import org.bitcoins.testkit.wallet.{BitcoinSWalletTest, WalletTestUtil}
 
 class AccountDAOTest extends BitcoinSWalletTest with AccountDAOFixture {
